@@ -2,11 +2,12 @@ package sandbox
 
 import (
 	"context"
+	"net/http"
+	"testing"
+
 	"github.com/gtforge/go-zooz"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"net/http"
-	"testing"
 )
 
 func TestCustomer(t *testing.T) {
